+++ src/FastAccelStepper.h
@@ -1,7 +1,7 @@
 #ifndef FASTACCELSTEPPER_H
 #define FASTACCELSTEPPER_H
 #include <stdint.h>
-#include "PoorManFloat.h"
+#include "Log2Representation.h"
 #include "fas_arch/common.h"
 
 // # FastAccelStepper