+++ src/RampGenerator.h
@@ -8,10 +8,10 @@
 
 class FastAccelStepper;
 
-#ifdef SUPPORT_PMF_TIMER_FREQ_VARIABLES
-extern pmf_logarithmic pmfl_timer_freq;
-extern pmf_logarithmic pmfl_timer_freq_div_sqrt_of_2;
-extern pmf_logarithmic pmfl_timer_freq_square_div_2;
+#ifdef SUPPORT_LOG2_TIMER_FREQ_VARIABLES
+extern pmf_logarithmic log2_timer_freq;
+extern pmf_logarithmic log2_timer_freq_div_sqrt_of_2;
+extern pmf_logarithmic log2_timer_freq_square_div_2;
 #endif
 
 class RampGenerator {